pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tangerine-meeting-assistant"
version = "1.6.0-beta.1"
description = "Your meeting -> your team's AI context, automatically."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Tangerine Intelligence" }]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.5",
    "pyyaml>=6",
    "rich>=13",
    "gitpython>=3.1",
    # Local Whisper transcription. CTranslate2-backed; runs on CPU with int8.
    # The bot's WHISPER_MODE=local path shells out to `python -m tmi.transcribe`,
    # which loads the model bundled by `tmi.model_download`.
    "faster-whisper>=1.0",
    # Used by tmi.model_download to fetch the small-int8 weights at first run.
    "huggingface_hub>=0.20",
]

[project.optional-dependencies]
dev = ["pytest>=7", "mypy>=1.8", "ruff>=0.4", "types-PyYAML"]

[project.scripts]
tmi = "tmi.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"tmi.observer" = ["prompts/*.md"]
"tmi" = ["prompts/*.txt"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
